OverridableMapping: implemented __getitem__ + tests
OverridableMapping differs from MergedMapping in that it permits to
completely override a key at any level in the hierarchy by appending a
colon to it. This is useful to let higher precedence scopes win over
lower ones (e.g. configure a system so that defaults can be overridden
at user level).