OverridableMapping now writes in an upper scratch layer
The strategy in OverridableMapping has been changed to write in an
upper scratch layer (in a similar way with respect to what is done
in the upper layer of a union filesystem).

The implementation details of MergedMapping and OverridableMapping
have been refactored to reduce code duplication. Tests have been
adapted to the new logic.