Refactoring sequences and mappings (#8)
Extracted a non-mutable version of merged sequence.
This permitted to fix the expected failures on tests.

Sequences and mappings have been moved into their own modules. Unit
tests have been refactored in the same way. For the time being names
are exported to the main package, to be readily usable.